django-service/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
